Move `Stage` class into `pypers.stage` and into `tests.test_stage`
