+++ pypers/pipeline.py
@@ -1,200 +1,17 @@
-import hashlib
-import json
 import os
-import re
-import time
 import weakref
 
-from pypers.config import Config
-from pypers.output import (
-    get_output,
-    Output,
-)
+from deprecated import deprecated
+
+import pypers.config
+import pypers.stage
+import pypers.status
 from pypers.typing import (
     Any,
-    Iterable,
     Optional,
     Sequence,
     Union,
 )
-
-
-def suggest_stage_id(class_name: str) -> str:
-    """
-    Suggest stage ID based on a class name.
-
-    This function validates the class name, then finds and groups tokens in the class name.
-    Tokens are grouped if they are consecutive and alphanumeric, but do not start with numbers.
-    The function then converts the tokens to lowercase, removes underscores, and joins them with hyphens.
-
-    :param class_name: The name of the class to suggest a configuration namespace for.
-    :type class_name: str
-    :return: A string of hyphen-separated tokens from the class name.
-    :rtype: str
-    :raises AssertionError: If the class name is not valid.
-    """
-    assert class_name != '_' and re.match('[a-zA-Z]', class_name) and re.match('^[a-zA-Z_](?:[a-zA-Z0-9_])*$', class_name), f'not a valid class name: "{class_name}"'
-    tokens1 = re.findall('[A-Z0-9][^A-Z0-9_]*', class_name)
-    tokens2 = list()
-    i1 = 0
-    while i1 < len(tokens1):
-        token = tokens1[i1]
-        i1 += 1
-        if len(token) == 1:
-            for t in tokens1[i1:]:
-                if len(t) == 1 and (token.isnumeric() == t.isnumeric() or token.isalpha() == t.isalpha()):
-                    token += t
-                    i1 += 1
-                else:
-                    break
-        tokens2.append(token.lower().replace('_', ''))
-    if len(tokens2) >= 2 and tokens2[-1] == 'stage': tokens2 = tokens2[:-1]
-    return '-'.join(tokens2)
-
-
-class Stage(object):
-    """
-    A pipeline stage.
-
-    Each stage can be controlled by a separate set of hyperparameters. Refer to the documentation of the respective pipeline stages for details. Most hyperparameters reside in namespaces, which are uniquely associated with the corresponding pipeline stages.
-
-    :param name: Readable identifier of this stage.
-    :param id: The stage ID, used as the hyperparameter namespace. Defaults to the result of the :py:meth:`~.suggest_stage_id` function if not specified.
-    :param inputs: List of inputs required by this stage.
-    :param outputs: List of outputs produced by this stage.
-
-    Automation
-    ^^^^^^^^^^
-
-    Hyperparameters can be set automatically using the :py:meth:`~.configure` method.
-
-    Inputs and outputs
-    ^^^^^^^^^^^^^^^^^^
-
-    Each stage must declare its required inputs and the outputs it produces. These are used by :py:meth:`~.create_pipeline` to automatically determine the stage order. The input ``input`` is provided by the pipeline itself.
-    """
-
-    inputs   = []
-    outputs  = []
-    consumes = []
-    enabled_by_default = True
-
-    def __init__(self):
-        self.id       = type(self).id if hasattr(type(self), 'id') else suggest_stage_id(type(self).__name__)
-        self.inputs   = frozenset(type(self).inputs) | frozenset(type(self).consumes)
-        self.outputs  = frozenset(type(self).outputs)
-        self.consumes = frozenset(type(self).consumes)
-        self.enabled_by_default = type(self).enabled_by_default
-        assert not self.id.endswith('+'), 'the suffix "+" is reserved as an indication of "the stage after that stage"'
-        self._callbacks = {}
-
-    def _callback(self, name, *args, **kwargs):
-        if name in self._callbacks:
-            for cb in self._callbacks[name]:
-                cb(self, name, *args, **kwargs)
-
-    def add_callback(self, name, cb):
-        if name == 'after':
-            self.add_callback( 'end', cb)
-            self.add_callback('skip', cb)
-        else:
-            if name not in self._callbacks: self._callbacks[name] = []
-            self._callbacks[name].append(cb)
-
-    def remove_callback(self, name, cb):
-        if name == 'after':
-            self.remove_callback( 'end', cb)
-            self.remove_callback('skip', cb)
-        else:
-            if name in self._callbacks: self._callbacks[name].remove(cb)
-
-    def __call__(self, data, cfg, out=None, log_root_dir=None, **kwargs):
-        out = get_output(out)
-        cfg = cfg.get(self.id, {})
-        if cfg.get('enabled', self.enabled_by_default):
-            out.intermediate(f'Starting stage "{self.id}"')
-            self._callback('start', data, out = out, **kwargs)
-            input_data = {key: data[key] for key in self.inputs}
-            clean_cfg = cfg.copy()
-            clean_cfg.pop('enabled', None)
-            t0 = time.time()
-            output_data = self.process(cfg=clean_cfg, log_root_dir=log_root_dir, out=out, **input_data)
-            dt = time.time() - t0
-            assert len(set(output_data.keys()) ^ set(self.outputs)) == 0, 'stage "%s" produced spurious or missing output' % self.id
-            data.update(output_data)
-            for key in self.consumes: del data[key]
-            self._callback('end', data, out = out, **kwargs)
-            return dt
-        else:
-            out.write(f'Skipping disabled stage "{self.id}"')
-            self._callback('skip', data, out = out, **kwargs)
-            return 0
-        
-    def skip(self, data, out = None, **kwargs):
-        self._callback('skip', data, out = out, **kwargs)
-
-    def process(self, cfg: Optional[Config]=None, log_root_dir: Optional[str]=None, out :Optional[Output]=None, **inputs):
-        """
-        Executes the current pipeline stage.
-
-        This method runs the current stage of the pipeline with the provided inputs, configuration parameters, and logging settings. It then returns the outputs produced by this stage.
-
-        :param input_data: A dictionary containing the inputs required by this stage. Each key-value pair in the dictionary represents an input name and its corresponding value.
-        :type input_data: dict
-        :param cfg: A dictionary containing the hyperparameters to be used by this stage. Each key-value pair in the dictionary represents a hyperparameter name and its corresponding value.
-        :type cfg: dict
-        :param log_root_dir: The path to the directory where log files will be written. If this parameter is ``None``, no log files will be written.
-        :type log_root_dir: str, optional
-        :param out: An instance of a subclass of :py:class:`~pypers.output.Output` to handle the output of this stage. If this parameter is ``'muted'``, no output will be produced. If this parameter is ``None``, the default output handler will be used.
-        :type out: :py:class:`~pypers.output.Output`, 'muted', or None, optional
-        :return: A dictionary containing the outputs produced by this stage. Each key-value pair in the dictionary represents an output name and its corresponding value.
-        :rtype: dict
-        """
-        raise NotImplementedError()
-
-    def configure(self, *args, **kwargs):
-        # FIXME: add documentation
-        return dict()
-    
-    @property
-    def signature(self):
-        signature = dict()
-
-        # Iterate over all attributes of the stage (leaving out a few special ones)
-        for key in dir(self):
-            if key in ('__doc__', '__weakref__', '__module__', '__dict__', 'signature', 'sha'): continue
-            value = getattr(self, key)
-
-            if isinstance(value, Iterable):
-                # Only keep the item if the iterable is is JSON-serializable
-                try:
-                    value = json.loads(json.dumps(list(value)))
-                except TypeError:
-                    continue
-
-            if callable(value):
-                # Only keep the item if has a custom implementation
-                try:
-                    value = value.__code__.co_code.hex()
-                except AttributeError:
-                    continue
-
-            # Add the item to the signature
-            signature[key] = value
-
-        # Return the signature
-        return signature
-
-    @property
-    def sha(self):
-        signature_str = json.dumps(self.signature)
-        return hashlib.sha1(signature_str.encode('utf-8')).hexdigest()
-
-    def __str__(self):
-        return self.id
-
-    def __repr__(self):
-        return f'<{type(self).__name__}, id: {self.id}>'
 
 
 class ProcessingControl:
@@ -277,7 +94,8 @@ class Configurator:
         """
         return self.pipeline.configure(base_cfg, input)
     
-    def first_differing_stage(self, config1: 'Config', config2: 'Config'):
+    @deprecated
+    def first_differing_stage(self, config1: pypers.config.Config, config2: pypers.config.Config):
         """
         Find the first stage with differing configurations between two sets of hyperparameters.
         
@@ -314,7 +132,7 @@ class Pipeline:
         self.stages = []
         self.configurator = configurator if configurator else Configurator(self)
 
-    def process(self, input, cfg, first_stage=None, last_stage=None, data=None, log_root_dir=None, out=None, **kwargs):  # TODO: Rename `cfg` to `config` and `input` to `file_id`
+    def process(self, input, cfg, first_stage=None, last_stage=None, data=None, log_root_dir=None, status=None, **kwargs):  # TODO: Rename `cfg` to `config` and `input` to `file_id`
         """
         Processes the input.
 
@@ -326,7 +144,7 @@ class Pipeline:
         :param last_stage: The name of the last stage to be executed.
         :param data: The results of a previous execution.
         :param log_root_dir: Path to a directory where log files should be written to.
-        :param out: An instance of an :py:class:`~pypers.output.Output` sub-class, ``'muted'`` if no output should be produced, or ``None`` if the default output should be used.
+        :param status: A :py:class:`~pypers.status.Status` object.
         :return: Tuple ``(data, cfg, timings)``, where ``data`` is the *pipeline data object* comprising all final and intermediate results, ``cfg`` are the finally used hyperparameters, and ``timings`` is a dictionary containing the execution time of each individual pipeline stage (in seconds).
 
         The parameter ``data`` is used if and only if ``first_stage`` is not ``None``. In this case, the outputs produced by the stages of the pipeline which are being skipped must be fed in using the ``data`` parameter obtained from a previous execution of this method.
@@ -340,19 +158,18 @@ class Pipeline:
         if data is None: data = dict()
         if input is not None: data['input'] = input
         extra_stages = self.get_extra_stages(first_stage, last_stage, data.keys())
-        out  = get_output(out)
         ctrl = ProcessingControl(first_stage, last_stage)
         timings = {}
         for stage in self.stages:
             if ctrl.step(stage.id) or stage.id in extra_stages:
                 try:
-                    dt = stage(data, cfg, out=out, log_root_dir=log_root_dir, **kwargs)
+                    dt = stage(data, cfg, status=pypers.status.derive(status), log_root_dir=log_root_dir, **kwargs)
                 except:
                     print(f'An error occured while executing the stage: {str(stage)}')
                     raise
                 timings[stage.id] = dt
             else:
-                stage.skip(data, out = out, **kwargs)
+                stage.skip(data, status = status, **kwargs)
         return data, cfg, timings
     
     def get_extra_stages(self, first_stage, last_stage, available_inputs):
@@ -374,7 +191,7 @@ class Pipeline:
             extra_stages.append(extra_stage.id)
         return extra_stages
 
-    def find(self, stage_id: str, not_found_dummy: Any = float('inf')) -> Stage:
+    def find(self, stage_id: str, not_found_dummy: Any = float('inf')) -> pypers.stage.Stage:
         """
         Returns the position of the stage identified by ``stage_id``.
 
@@ -389,7 +206,7 @@ class Pipeline:
         idx = self.find(stage_id, None)
         return self.stages[idx] if idx is not None else None
 
-    def append(self, stage: 'Stage', after: Union[str, int] = None):
+    def append(self, stage: pypers.stage.Stage, after: Union[str, int] = None):
         for stage2 in self.stages:
             if stage2 is stage: raise RuntimeError(f'stage {stage.id} already added')
             if stage2.id == stage.id: raise RuntimeError(f'stage with ID {stage.id} already added')
@@ -423,7 +240,7 @@ class Pipeline:
         return frozenset(fields)
 
 
-def create_pipeline(stages: Sequence[Stage], *args, **kwargs) -> Pipeline:
+def create_pipeline(stages: Sequence[pypers.stage.Stage], *args, **kwargs) -> Pipeline:
     """
     Creates and returns a new :py:class:`.Pipeline` object configured for the given stages.
