+++ tests/test_pipeline.py
@@ -1,5 +1,6 @@
 import itertools
 import unittest
+from unittest.mock import MagicMock
 
 import pypers.pipeline
 import pypers.config
@@ -7,120 +8,9 @@ import pypers.config
 from . import testsuite
 
 
-class suggest_id(unittest.TestCase):
-
-    def test(self):
-        self.assertEqual(pypers.pipeline.suggest_stage_id('TheGreatPCMapper'     ), 'the-great-pc-mapper'    )
-        self.assertEqual(pypers.pipeline.suggest_stage_id('TheGreat_PCMapper'    ), 'the-great-pc-mapper'    )
-        self.assertEqual(pypers.pipeline.suggest_stage_id('TheGreat__PCMapper'   ), 'the-great-pc-mapper'    )
-        self.assertEqual(pypers.pipeline.suggest_stage_id('TheGreat_123_PCMapper'), 'the-great-123-pc-mapper')
-        self.assertEqual(pypers.pipeline.suggest_stage_id('TheGreat123_PCMapper' ), 'the-great-123-pc-mapper')
-        self.assertEqual(pypers.pipeline.suggest_stage_id('TheGreat123PCMapper'  ), 'the-great-123-pc-mapper')
-        self.assertEqual(pypers.pipeline.suggest_stage_id('TheGreatMapperStage'  ), 'the-great-mapper'       )
-        self.assertEqual(pypers.pipeline.suggest_stage_id('Stage'                ), 'stage'                  )
-
-    def test_illegal(self):
-        self.assertRaises(AssertionError, lambda: pypers.pipeline.suggest_stage_id(''))
-        self.assertRaises(AssertionError, lambda: pypers.pipeline.suggest_stage_id('_'))
-        self.assertRaises(AssertionError, lambda: pypers.pipeline.suggest_stage_id('_1'))
-        self.assertRaises(AssertionError, lambda: pypers.pipeline.suggest_stage_id('TheGreat PCMapper'))
-        self.assertRaises(AssertionError, lambda: pypers.pipeline.suggest_stage_id('TheGreat-PCMapper'))
-        self.assertRaises(AssertionError, lambda: pypers.pipeline.suggest_stage_id('1TheGreatPCMapper'))
-
-
-class Stage(unittest.TestCase):
-
-    def test_no_inputs_no_outputs(self):
-        stage = testsuite.create_stage(id = 'test')
-        data  = dict()
-        cfg   = pypers.config.Config()
-        dt    = stage(data, cfg, out = 'muted')
-        self.assertIsInstance(dt, float)
-        self.assertEqual(data, dict())
-
-    def test_init(self):
-        class Stage(pypers.pipeline.Stage):
-            pass
-        self.assertEqual(Stage().id, 'stage')
-
-    def test(self):
-        stage = testsuite.create_stage(id = 'test', inputs = ['x1', 'x2'], outputs = ['y'], \
-            process = lambda x1, x2, cfg, log_root_dir = None, out = None: \
-                dict(y = \
-                    x1 * cfg.get('x1_factor', 0) + \
-                    x2 * cfg.get('x2_factor', 0))
-            )
-        cfg = pypers.config.Config()
-        for x1_factor in [0, 1]:
-            for x2_factor in [0, 1]:
-                x1, x2 = 10, 20
-                with self.subTest(x1_factor = x1_factor, x2_factor = x2_factor):
-                    cfg['test/x1_factor'] = x1_factor
-                    cfg['test/x2_factor'] = x2_factor
-                    data = dict(x1 = x1, x2 = x2)
-                    dt   = stage(data, cfg, out = 'muted')
-                    self.assertEqual(data, dict(x1 = x1, x2 = x2, y = x1 * x1_factor + x2 * x2_factor))
-                    self.assertIsInstance(dt, float)
-
-    def test_missing_input(self):
-        stage = testsuite.create_stage(id = 'test', outputs = ['y'], \
-            process = lambda x, cfg, log_root_dir = None, out = None: \
-                dict(y = x)
-            )
-        data = dict(x = 0)
-        cfg  = pypers.config.Config()
-        self.assertRaises(TypeError, lambda: stage(data, cfg, out = 'muted'))
-
-    def test_missing_output(self):
-        stage = testsuite.create_stage(id = 'test', outputs = ['y'], \
-            process = lambda cfg, log_root_dir = None, out = None: \
-                dict()
-            )
-        data = dict()
-        cfg  = pypers.config.Config()
-        self.assertRaises(AssertionError, lambda: stage(data, cfg, out = 'muted'))
-
-    def test_spurious_output(self):
-        stage = testsuite.create_stage( id = 'test', \
-            process = lambda cfg, log_root_dir = None, out = None: \
-                dict(y = 0)
-            )
-        data = dict()
-        cfg  = pypers.config.Config()
-        self.assertRaises(AssertionError, lambda: stage(data, cfg, out = 'muted'))
-
-    def test_missing_and_spurious_output(self):
-        stage = testsuite.create_stage(id = 'test', outputs = ['y'], \
-            process = lambda cfg, log_root_dir = None, out = None: \
-                dict(z = 0)
-            )
-        data = dict()
-        cfg  = pypers.config.Config()
-        self.assertRaises(AssertionError, lambda: stage(data, cfg, out = 'muted'))
-
-    def test_consumes(self):
-        stage = testsuite.create_stage(id = 'test', consumes = ['x'], \
-            process = lambda x, cfg, log_root_dir = None, out = None: \
-                dict()
-            )
-        data = dict(x = 0, y = 1)
-        cfg  = pypers.config.Config()
-        stage(data, cfg, out = 'muted')
-        self.assertEqual(data, dict(y = 1))
-
-    def test_missing_consumes(self):
-        stage = testsuite.create_stage(id = 'test', consumes = ['x'], \
-            process = lambda x, cfg, log_root_dir = None, out = None: \
-                dict()
-            )
-        data = dict()
-        cfg  = pypers.config.Config()
-        self.assertRaises(KeyError, lambda: stage(data, cfg, out = 'muted'))
-
-
 class Pipeline(unittest.TestCase):
 
-    def _test_configure(self, original_base_cfg: 'config.Config'):
+    def _test_configure(self, original_base_cfg: pypers.config.Config):
         factors = list(range(-1, +3))
         for x1_pre_factor, x2_pre_factor in itertools.product(factors, factors):
             for x1_default_user_factor, x2_default_user_factor in itertools.product(factors, factors):
@@ -226,7 +116,7 @@ class Pipeline(unittest.TestCase):
         cfg['stage3/constant' ] = 3
         pipeline = create_pipeline().test()
         for stage in pipeline.stages:
-            def _cb(stage2, name, data, out):
+            def _cb(stage2, name, data, status):
                 pipeline.call_record.append(f'{stage2.id} {name}')
             stage.add_callback('start', _cb)
             stage.add_callback('end'  , _cb)
@@ -241,13 +131,15 @@ class Pipeline(unittest.TestCase):
             ]) for stage in pipeline.stages), [])
         for suffix, offset in (('', 0), ('+', 1)):
             for input in range(3):
-                full_data, _, _ = pipeline.process(input = input, cfg = cfg, out = 'muted')
+                status_mock = MagicMock()
+                full_data, _, _ = pipeline.process(input = input, cfg = cfg, status = status_mock)
                 self.assertEqual(pipeline.call_record, expected_call_record([stage.id for stage in pipeline.stages]))
                 pipeline.call_record.clear()
                 for first_stage_idx, first_stage in enumerate((stage.id for stage in pipeline.stages[:len(pipeline.stages) - offset])):
                     with self.subTest(suffix = suffix, offset = offset, input = input, first_stage = first_stage):
                         remaining_stages = frozenset([stage.id for stage in pipeline.stages[first_stage_idx + offset:]])
-                        data, _, timings = pipeline.process(input = input, data = full_data, first_stage = first_stage + suffix, cfg = cfg, out = 'muted')
+                        status_mock = MagicMock()
+                        data, _, timings = pipeline.process(input = input, data = full_data, first_stage = first_stage + suffix, cfg = cfg, status = status_mock)
                         self.assertEqual(data['y'], full_data['y'])
                         self.assertEqual(frozenset(timings.keys()), remaining_stages)
                         self.assertEqual(pipeline.call_record, expected_call_record(remaining_stages))
@@ -275,15 +167,15 @@ class create_pipeline(unittest.TestCase):
 
     def test(self, configure_stage1 = None, configure_stage2 = None, configure_stage3 = None):
         call_record = list()
-        def _process_stage1(input, cfg, log_root_dir = None, out = None):
+        def _process_stage1(input, cfg, log_root_dir = None, status = None):
             call_record.append('stage1 process')
             return dict(x1 = input * cfg['x1_factor'])
         
-        def _process_stage2(input, cfg, log_root_dir = None, out = None):
+        def _process_stage2(input, cfg, log_root_dir = None, status = None):
             call_record.append('stage2 process')
             return dict(x2 = input * cfg['x2_factor'])
         
-        def _process_stage3(x1, x2, cfg, log_root_dir = None, out = None):
+        def _process_stage3(x1, x2, cfg, log_root_dir = None, status = None):
             call_record.append('stage3 process')
             return dict(y = x1 + x2 + cfg['constant'])
         stages = [
@@ -335,7 +227,3 @@ class create_pipeline(unittest.TestCase):
         stage2 = testsuite.create_stage(id = 'stage2', outputs = ['x2'])
         stage3 = testsuite.create_stage(id = 'stage3', outputs = ['x2'])
         self.assertRaises(AssertionError, lambda: pypers.pipeline.create_pipeline([stage1, stage2, stage3]))
-
-
-if __name__ == '__main__':
-    unittest.main()
